+++ src/components/index.ts
@@ -1,7 +1,7 @@
 export {AccommodationCard} from "./Cards/AccommodationCard"
 export {AccommodationSearch} from "./Search/AccommodationSearch"
 export {AdvancedSearch} from "./Search/AdvancedSearch"
-export {BookStayButton} from "./BookStayButton"
+export {BookStayButton} from "./Buttons/BookStayButton"
 export {CityCard} from "./Cards/CityCard"
 export {Footer} from "./Footer"
 export {Header} from "./Header"
@@ -10,5 +10,5 @@ export {PlaceCard} from "./Cards/PlaceCard"
 export {Rating} from "./Rating"
 export {ReservationCard} from "./Cards/ReservationCard"
 export {ReservationForm} from "./ReservationForm"
-export {SearchButton} from "./SearchButton"
+export {SearchButton} from "./Buttons/SearchButton"
 export {SimpleSearch} from "./Search/SimpleSearch"
